# EduCloud-v42/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EduCloud API v1",
    "version": "v1"
  },
  "paths": {
    "/api/CourseElementsApi": {
      "get": {
        "tags": [
          "CourseElementsApi"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseElement"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseElement"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseElement"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CourseElementsApi"
        ],
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseElement"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseElement"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CourseElement"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CourseElement"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseElement"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseElement"
                }
              }
            }
          }
        }
      }
    },
    "/api/CourseElementsApi/{id}": {
      "get": {
        "tags": [
          "CourseElementsApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CourseElement"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseElement"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseElement"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CourseElementsApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseElement"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseElement"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CourseElement"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "CourseElementsApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/courses": {
      "get": {
        "tags": [
          "CoursesApi"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CoursesApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/courses/{id}": {
      "get": {
        "tags": [
          "CoursesApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CoursesApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "CoursesApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/UsersApi": {
      "get": {
        "tags": [
          "UsersApi"
        ],
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "UsersApi"
        ],
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/UsersApi/{id}": {
      "get": {
        "tags": [
          "UsersApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "UsersApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "UsersApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "userCourses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserCourse"
            },
            "nullable": true
          },
          "courseElements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseElement"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CourseElement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/CourseElementType"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "courseId": {
            "type": "integer",
            "format": "int32"
          },
          "course": {
            "$ref": "#/components/schemas/Course"
          },
          "courseFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseFile"
            },
            "nullable": true
          },
          "userTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserTask"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CourseElementType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "CourseFile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "courseElementId": {
            "type": "integer",
            "format": "int32"
          },
          "courseElement": {
            "$ref": "#/components/schemas/CourseElement"
          }
        },
        "additionalProperties": false
      },
      "CourseRole": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "TaskFile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "taskId": {
            "type": "integer",
            "format": "int32"
          },
          "userTask": {
            "$ref": "#/components/schemas/UserTask"
          }
        },
        "additionalProperties": false
      },
      "User": {
        "required": [
          "email",
          "fullName",
          "passwordHash",
          "username"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "fullName": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "email": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "format": "tel",
            "nullable": true
          },
          "passwordHash": {
            "minLength": 1,
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "userCourses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserCourse"
            },
            "nullable": true
          },
          "userTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserTask"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserCourse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "courseId": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "$ref": "#/components/schemas/CourseRole"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "course": {
            "$ref": "#/components/schemas/Course"
          }
        },
        "additionalProperties": false
      },
      "UserRole": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "UserTask": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "taskId": {
            "type": "integer",
            "format": "int32"
          },
          "mark": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "task": {
            "$ref": "#/components/schemas/CourseElement"
          },
          "taskFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskFile"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}